refactor(bindings): cleanup nodejs binding
