+++ bindings/node/native/src/classes/mod.rs
@@ -1,0 +1,7 @@
+mod account;
+mod account_manager;
+mod synced_account;
+
+pub use account::*;
+pub use account_manager::*;
+pub use synced_account::*;
